fast-api-sqlite-demo/requirements.txt
# fast api basic
fastapi~=0.109.2
uvicorn[standard]~=0.27.1
# seems to be included in fast api might be not relvant
pydantic~=2.6.2
# sql database
SQLAlchemy~=2.0.27
# auth
passlib[bcrypt]~=1.7.4
bcrypt~=4.1.2
python-multipart~=0.0.9
python-jose[cryptography]~=3.3.0
